Tout semble fonctionner sur Biblio_GAUTIER>src>biblio>view>Ecran_Acceuil

+++ Biblio_GAUTIER/src/biblio/dao/EmpruntEnCoursDao.java
@@ -115,36 +115,36 @@ public class EmpruntEnCoursDao {
 		Statement stmt = cnx3.createStatement();
 		
 		ResultSet rs8 = stmt.executeQuery("SELECT * FROM EMPRUNTENCOURS WHERE IDEXEMPLAIRE ="+ idExemplaire);
-		rs8.next();
+		if(rs8.next()) {
 		
-		Statement stmt2 = cnx3.createStatement();
-		
-		ResultSet rs3 = stmt2.executeQuery("SELECT IDEMPRUNTARCHIVE+1 FROM EMPRUNTARCHIVE ORDER BY IDEMPRUNTARCHIVE DESC FETCH FIRST 1 ROWS ONLY");
-		rs3.next();
-		
-		
-		PreparedStatement pstmt6 = cnx3.prepareStatement("INSERT INTO EMPRUNTARCHIVE VALUES (?, TO_DATE(?, 'DD-MM-YYYY'), TO_DATE(?, 'DD-MM-YYYY'),?,?)");
-		pstmt6.setInt(1,(rs3.getInt(1)));
-		pstmt6.setDate(2,rs8.getDate(3));
-		pstmt6.setDate(3,Date.valueOf(LocalDate.now()));
-		pstmt6.setInt(4,rs8.getInt(1));
-		pstmt6.setInt(5,rs8.getInt(2));
-		pstmt6.executeUpdate();
-		pstmt6.close();
-		rs8.close();
-		rs3.close();
-		stmt.close();
-		stmt2.close();
-		
-		PreparedStatement pstmt2 = cnx3.prepareStatement("DELETE FROM EMPRUNTENCOURS WHERE IDEXEMPLAIRE = ?");
-		pstmt2.setInt(1,idExemplaire);
-		pstmt2.execute();
-		pstmt2.close();
-		PreparedStatement pstmt3 = cnx3.prepareStatement("UPDATE EXEMPLAIRE SET STATUS='DISPONIBLE' WHERE IDEXEMPLAIRE = ?");
-		pstmt3.setInt(1,idExemplaire);
-		pstmt3.executeUpdate();
-		pstmt3.close();	
-		
+			Statement stmt2 = cnx3.createStatement();
+			
+			ResultSet rs3 = stmt2.executeQuery("SELECT IDEMPRUNTARCHIVE+1 FROM EMPRUNTARCHIVE ORDER BY IDEMPRUNTARCHIVE DESC FETCH FIRST 1 ROWS ONLY");
+			rs3.next();
+			
+			
+			PreparedStatement pstmt6 = cnx3.prepareStatement("INSERT INTO EMPRUNTARCHIVE VALUES (?, TO_DATE(?, 'DD-MM-YYYY'), TO_DATE(?, 'DD-MM-YYYY'),?,?)");
+			pstmt6.setInt(1,(rs3.getInt(1)));
+			pstmt6.setDate(2,rs8.getDate(3));
+			pstmt6.setDate(3,Date.valueOf(LocalDate.now()));
+			pstmt6.setInt(4,rs8.getInt(1));
+			pstmt6.setInt(5,rs8.getInt(2));
+			pstmt6.executeUpdate();
+			pstmt6.close();
+			rs8.close();
+			rs3.close();
+			stmt.close();
+			stmt2.close();
+			
+			PreparedStatement pstmt2 = cnx3.prepareStatement("DELETE FROM EMPRUNTENCOURS WHERE IDEXEMPLAIRE = ?");
+			pstmt2.setInt(1,idExemplaire);
+			pstmt2.execute();
+			pstmt2.close();
+			PreparedStatement pstmt3 = cnx3.prepareStatement("UPDATE EXEMPLAIRE SET STATUS='DISPONIBLE' WHERE IDEXEMPLAIRE = ?");
+			pstmt3.setInt(1,idExemplaire);
+			pstmt3.executeUpdate();
+			pstmt3.close();	
+			}
 		cnx3.close();
 			
 //		int j= 0;
